requirements.txt
django-heroku
gunicorn
psycopg2
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
confusable-homoglyphs==3.2.0
cycler==0.10.0
django-bootstrap-form==3.4
django-registration==3.1
django-tastypie==0.14.3
idna==2.10
kiwisolver==1.2.0
matplotlib==3.3.1
numpy==1.19.1
ordereddict==1.1
pandas==1.1.1
Pillow==7.2.0
polib==1.1.0
pyparsing==2.4.7
pySankeyBeta==1.2.2
python-dateutil==2.8.1
python-mimeparse==1.6.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
scipy==1.5.2
seaborn==0.10.1
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
whitenoise==5.2.0